% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcd.R
\name{crcd}
\alias{crcd}
\title{Calculate conditional robust copula dependence}
\usage{
crcd(x, y, z, density = "kde", k, bandwidth, cpp = "parallel",
  verbose = F)
}
\arguments{
\item{x}{The vector for the first random variable}

\item{y}{The vector for the second random variable}

\item{z}{The vector for the third random variable that conditioned on}

\item{density}{The method used in density estimation, i.e. knn or kde. This parameter could be missing and a default value will be applied.}

\item{k}{The parameter K in KNN density estimation. This parameter could be missing and a default value will be applied.}

\item{bandwidth}{The bandwidth used in the density estimation. This parameter could be missing and a default value will be applied.}

\item{cpp}{The ways to calculate the distance matrix. For KDE, cpp=FALSE means using pure R implementation, otherwise, using c++.For KNN, "none" means using R distance function, "serial" means using serial version of Rcpp, "parallel" means using parallel version of Rcpp. This parameter could be missing and a default value will be applied.}

\item{verbose}{Print the current case or not. Default is False.}
}
\value{
The RCD of \code{x} and \code{y} given \code{z}
}
\description{
This is the function that used to calculate the conditional robust copula dependence (RCD)
between two random variables \code{x} and \code{y} given \code{z}. Note that the length of
\code{x} \code{y} and \code{z} should be the same.
}
\examples{
n <- 1000
 x <- runif(n)
 y <- runif(n)
 rcd(x,x,y)
}

