% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over.R
\name{over}
\alias{over}
\title{over
Calculates regulated cell death (rcd) enrichment on the input custom gene list}
\usage{
custom<-c("CALR","KLRK1","HMGB1","HSPA1A","HSPA1B","HSPA1L","FLNA","FLNB","MYH9","TLN1","ACTB","MYL6","MYH10")

load(rcd)

res<-over(custom,rcd)

res
}
\arguments{
\item{custom}{a vector of gene symbol containing possible rcd molecules}
}
\description{
over
Calculates regulated cell death (rcd) enrichment on the input custom gene list
}
\examples{
load(rcd)
custom<-c("CALR","KLRK1","HMGB1","HSPA1A","HSPA1B","HSPA1L","FLNA","FLNB","MYH9","TLN1","ACTB","MYL6","MYH10")
res<-over(custom,rcd)
res
}
