\name{pharm}
\alias{pharm}

\title{
Pharmaceutical company stock price .
}

\description{
The stock prices for three pharmaceutical companies and a market index.
}

\usage{data(pharm)}

\details{
	This data set contains 4054 samples with 4 predictors
}

\format{
  The format is a list containing conatins a dataframer.
  1. pharm -stock prices for three pharmaceutical companies and a market index.
}


\seealso{
\code{\link{http://finance.yahoo.com/}}.
}

\examples{
data(pharm)
pairs(pharm)
}
