% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotscatter.R
\name{plotscatter}
\alias{plotscatter}
\title{plotscatter
perform scatterplot of the rcd enrichment}
\usage{
custom<-c("CALR","KLRK1","HMGB1","HSPA1A","HSPA1B","HSPA1L","FLNA","FLNB","MYH9","TLN1","ACTB","MYL6","MYH10")

load(rcd)

res<-over(custom,rcd)

plotscatter(res,x_min=-10,x_max=70,y_min=0,y_max=10,label=5,base=18)
}
\arguments{
\item{res}{results of the rcd enrichment obtained by over function}

\item{x_min}{minimum for x axis}

\item{x_max}{maximum for x axis}

\item{y_min}{minimum for y axis}

\item{y_max}{maximum for y axis}

\item{label}{size of the labels}

\item{base}{base size of the ggplot}
}
\description{
plotscatter
perform scatterplot of the rcd enrichment
}
\examples{
load(rcd)
custom<-c("CALR","KLRK1","HMGB1","HSPA1A","HSPA1B","HSPA1L","FLNA","FLNB","MYH9","TLN1","ACTB","MYL6","MYH10")
res<-over(custom,rcd)
res
plotscatter(res,x_min=-30,x_max=70,y_min=0,y_max=10,label=5,base=18)
}
