% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcd.R
\name{rcd}
\alias{rcd}
\title{Calculate robust copula dependence}
\usage{
rcd(x, y, method = "knn", k, bw, typ = 0, cpp = 2)
}
\arguments{
\item{x}{The sampled data, a vector, for the first random variable}

\item{y}{The sampled data, a vector, for the second random variable}

\item{method}{The method used in density estimation, i.e. knn or kde.}

\item{k}{The parameter K in KNN that need to be used in the density estimation. This parameter could be missing and a deault value will be applied.}

\item{bw}{The bandwidth need to be used in the density estimation. This parameter could be missing and a deault value will be applied.}

\item{typ}{The ways to calculate the density. "0" means the density is calculated by using the result in the paper Kraskov(2003). "1" means the density is calculated by using the paper Loftsgaarden and Quesenberry (1965).}

\item{cpp}{The ways to calculate the distance matrix. "0" means using R distance function, "1" means using serial version of Rcpp, "2" means using parallel version of Rcpp.}
}
\value{
The RCD of \code{x} and \code{y}
}
\description{
This is the function that used to calculate the robust copula dependence (RCD)
between two random variables \code{x} and \code{y}. Note that the length of
\code{x} and \code{y} should be the same.
}
\examples{
n <- 1000
 x <- runif(n)
 y <- x + 0.5 * runif(n)
 rcd(x,y,method="kde")
}

