% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdheat.R
\name{rcdheat}
\alias{rcdheat}
\title{rcdheat}
\usage{
data(matrix)

data(rcd)

data(phenotype)

ssres<-rcdscore(expr,rcd,method="zscore")

rcdheat(ssres,phenotype,scale="none",fontsize=10)
}
\arguments{
\item{ssres}{results obtained from rcdscore}

\item{phenotype}{phenotype file from tested samples}

\item{scale}{method of scaling "none" "row" "column"}

\item{fontsize}{size of the font on the graph}
}
\description{
rcdheat
}
\examples{
data(matrix)
data(rcd)
data(phenotype)
ssres<-rcdscore(expr,rcd,method="zscore")
rcdheat(ssres,phenotype,scale="none",fontsize=10)
}
