% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcdknn.R
\name{rcdknn}
\alias{rcdknn}
\title{Calculate robust copula dependence}
\usage{
rcdknn(x, y, cc = 1/4, typ = 0, cpp = 2)
}
\arguments{
\item{x}{The sampled data, a vector, for the first random variable}

\item{y}{The sampled data, a vector, for the second random variable}

\item{cc}{The bandwidth need to be used in the density estimation. This parameter could be missing and a deault value will be applied.}

\item{typ}{The ways to calculate the density}

\item{cpp}{Using cpp or not}
}
\value{
The RCD of \code{x} and \code{y}
}
\description{
This is the function that used to calculate the robust copula dependence (RCD)
between two random variables \code{x} and \code{y}. Note that the length of
\code{x} and \code{y} should be the same.
}
\examples{
n <- 1000
 x <- runif(n)
 y <- runif(n)
 rcdknn(x,y)
}

