% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdpca.R
\name{rcdpca}
\alias{rcdpca}
\title{rcdpca}
\usage{
data(matrix)

data(rcd)

data(phenotype)

ssres<-rcdscore(expr,rcd,method="zscore")

rcdpca(ssres,phenotype,group="condition",scale=TRUE,fontsize=16)
}
\arguments{
\item{ssres}{results obtained from rcdscore}

\item{phenotype}{phenotype file from tested samples}

\item{fontsize}{size of the font on the graph}

\item{scale}{scaling TRUE or FALSE}

\item{alpha}{dot transparency}
}
\description{
rcdpca
}
\examples{
data(matrix)
data(rcd)
data(phenotype)
ssres<-rcdscore(expr,rcd,method="zscore")
rcdpca(ssres,phenotype,group="condition",scale=TRUE,fontsize=16)
}
