% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdscore.R
\name{rcdscore}
\alias{rcdscore}
\title{rcdscore}
\usage{
data(matrix)

data(rcd)

ssres<-rcdscore(expr,rcd,method="zscore")

head(ssres[,1:6])
}
\arguments{
\item{expr}{transcriptome matrix}

\item{rcd}{human genesets of rcd}

\item{method}{GSVA method of enrichment: "gsva", "ssgsea", "zscore", "plage"}

\item{rcdmm}{mus musculus genesets of rcd}
}
\description{
rcdscore
}
\examples{
data(matrix)
data(rcd)
ssres<-rcdscore(expr,rcd,method="zscore")
head(ssres[,1:6])
}
