% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcd.R
\name{rcd}
\alias{rcd}
\title{Calculate robust copula dependence}
\usage{
rcd(x, y, method = "knn", k, bandwidth, M = 200, cpp = "parallel")
}
\arguments{
\item{x}{The vector for the first random variable}

\item{y}{The vector for the first random variable}

\item{method}{The method used in density estimation, i.e. knn or kde.}

\item{k}{The parameter K in KNN density estimation. This parameter could be missing and a default value will be applied.}

\item{bandwidth}{The bandwidth used in the density estimation. This parameter could be missing and a default value will be applied.}

\item{M}{Number of steps used in numerical integration. This parameter could be missing and a default value will be applied.}

\item{cpp}{The ways to calculate the distance matrix. "none" means using R distance function, "serial" means using serial version of Rcpp, "parallel" means using parallel version of Rcpp. This parameter could be missing and a default value will be applied.}
}
\value{
The RCD of \code{x} and \code{y}
}
\description{
This is the function used to calculate the robust copula dependence (RCD)
between two random variables \code{x} and \code{y}. Note that the length of
\code{x} and \code{y} should be the same.
}
\examples{
n <- 1000
 x <- runif(n)
 y <- x^2 + 2*runif(n)
 rcd(x,y,method="kde")
}

