% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcdkde.R
\name{rcdkde}
\alias{rcdkde}
\title{Calculate robust copula dependence with the KDE estimator}
\usage{
rcdkde(x, y, bandwidth, M)
}
\arguments{
\item{x}{The vector for the first random variable}

\item{y}{The vector for the second random variable}

\item{bandwidth}{The bandwidth used in the density estimation. This parameter could be missing and a default value will be applied.}

\item{M}{Number of steps used in numerical integration. This parameter could be missing and a default value will be applied.}
}
\value{
The RCD of \code{x} and \code{y}
}
\description{
This is the function used to calculate the robust copula dependence (RCD)
between two random variables \code{x} and \code{y} with the KDE estimator. Note that the length of
\code{x} and \code{y} should be the same.
}
\examples{
n <- 1000
 x <- runif(n)
 y <- x^2 + 2*runif(n)
 rcdkde(x,y)
}

