% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcdknn.R
\name{rcdknn}
\alias{rcdknn}
\title{Calculate robust copula dependence with the KNN estimator}
\usage{
rcdknn(x, y, k, cpp = "parallel")
}
\arguments{
\item{x}{The vector for the first random variable}

\item{y}{The vector for the second random variable}

\item{k}{The parameter K in KNN that need to be used in the density estimation. This parameter could be missing and a deault value will be applied.}

\item{cpp}{The ways to calculate the distance matrix. "none" means using R distance function, "serial" means using serial version of Rcpp, "parallel" means using parallel version of Rcpp. This parameter could be missing and a default value will be applied.}
}
\value{
The RCD of \code{x} and \code{y}
}
\description{
This is the function used to calculate the robust copula dependence (RCD)
between two random variables \code{x} and \code{y} with the KNN estimator. Note that the length of
\code{x} and \code{y} should be the same.
}
\examples{
n <- 1000
 x <- runif(n)
 y <- x^2 + 2*runif(n)
 rcdknn(x,y)
}

